+++ docs/changelog.md
@@ -2,6 +2,7 @@
 
 ## [0.2.6] - Unreleased
 - Added `KY` and `MM` to valid prefixes
+- Added support to check for compiled extension modules
 
 ## [0.2.5] - 2022-04-10
 - Fixed missing `jcs` dependency